Replace JavaScript with TypeScript
